adding test for the success/error handler
