+++ libs/clients/driving-license/src/lib/drivingLicenseApi.service.ts
@@ -3,6 +3,7 @@ import { DrivingAssessment, Juristiction, QualityPhoto } from '..'
 import { ApiV1, EmbaettiDto } from '../v1'
 import { ApiV2, DRIVING_LICENSE_API_VERSION_V2, Rettindi } from '../v2'
 import { DriversLicense, Teacher } from './drivingLicenseApi.types'
+import { handleCreateResponse } from './utils/handleCreateResponse'
 
 // empty string === successful license posted!?!
 const DRIVING_LICENSE_SUCCESSFUL_RESPONSE_VALUE = ''
@@ -209,7 +210,7 @@ export class DrivingLicenseApi {
       apiVersion: DRIVING_LICENSE_API_VERSION_V2,
     })
 
-    const handledResponse = this.getCreateResponse(response)
+    const handledResponse = handleCreateResponse(response)
 
     if (!handledResponse.success) {
       throw new Error(
@@ -218,33 +219,6 @@ export class DrivingLicenseApi {
     }
 
     return handledResponse.success
-  }
-
-  // service returns response in the form of { value: number } but openapi doc says
-  // that it only returns a number - and it returns a string on error
-  private getCreateResponse(
-    response: unknown,
-  ): { error?: string | unknown; success: boolean } {
-    if (response === '') {
-      return { success: true }
-    }
-
-    try {
-      const resObj = JSON.parse(response as string)
-
-      if (resObj?.value) {
-        return { success: true }
-      } else if (typeof resObj === 'number') {
-        // if it's a number it means the API now returns the same type as is documented
-        return { success: true }
-      } else if (typeof resObj === 'string') {
-        return { success: false, error: resObj }
-      } else {
-        return { success: false, error: 'unknown type of response' }
-      }
-    } catch (e) {
-      return { success: false, error: e }
-    }
   }
 
   async getHasQualityPhoto(params: { nationalId: string }): Promise<boolean> {
